host/requirements.txt
# SpeechMate Host Server Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
flask==3.0.0
werkzeug==3.0.1

# ASR Models
faster-whisper==1.0.1

# Audio Processing
soundfile==0.12.1

# Database
sqlalchemy==2.0.25

# HTTP Client
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart==0.0.6

# Logging
loguru==0.7.2

# System
psutil==5.9.7

# Optional: Local Translation Models (uncomment if needed)
# transformers==4.37.0
# torch==2.1.2
# sentencepiece==0.1.99
# sacremoses==0.1.1
